+++ sdk/release_notes.md
@@ -12,6 +12,7 @@
 
 - Add analyzer for SupportsDeferredBindingAttribute #1367
 - Added an analyzer that will show a warning for types not supported by a binding attribute (#1505)
+- Added an analyzer that will suggest a code refactor for all of the types supported by a binding attribute (#1604)
 
 ### Microsoft.Azure.Functions.Worker.Sdk.Generators <version>